requirements.txt
fastapi[standard]
sqlmodel
python-dotenv
